only install llama when currently installed is not the same

+++ pkg/llama-cpp-vulkan-bin.sh
@@ -3,8 +3,16 @@ set -eo pipefail -ux
 
 pushd "${BASH_SOURCE%/*}"/llama-cpp-vulkan-bin
 
-makepkg --force --install --noconfirm --syncdeps
-git clean -dfX
-git checkout PKGBUILD
+# shellcheck disable=SC1091
+AVAILABLE=$(source PKGBUILD && pkgver)
+INSTALLED=$(pacman -Q llama-cpp-vulkan-bin 2> /dev/null | awk '{print $2}' | cut -d- -f1)
+
+if [[ $AVAILABLE != "$INSTALLED" ]]; then
+
+  makepkg --force --install --noconfirm --syncdeps
+  git clean -dfX
+  git checkout PKGBUILD
+
+fi
 
 popd
